--- app/news/news.js
@@ -1,8 +1,0 @@
-
-const news = [
-    { id: 1, name: 'news1' },
-    { id: 2, name: 'news2' },
-    { id: 3, name: 'news3' }
-]
-
-module.exports = news;
