+++ src/app/news/delete.js
@@ -1,8 +1,7 @@
 const express = require('express');
 const router = express.Router();
-const news = require('./news');
 
-router.delete('/:id', (req, res) => {
+module.exports = router.delete('/:id', (req, res) => {
     const newsData = news.find(c => c.id === parseInt(req.params.id));
     if (!newsData) return res.status(404).send('The news with given Id was not found');
     const index = news.indexOf(newsData);
@@ -10,4 +9,8 @@ router.delete('/:id', (req, res) => {
     res.send(newsData);
 });
 
-module.exports = router;
+const news = [
+    { id: 1, name: 'news1' },
+    { id: 2, name: 'news2' },
+    { id: 3, name: 'news3' }
+]
