+++ src/app/news/index.js
@@ -1,0 +1,13 @@
+// const get = require('./get')
+const post = require('./post')
+const get = require('./get')
+// const put = require('./put')
+// const remove = require('./delete')
+
+module.exports = {
+  get,
+  post
+//   put,
+//   remove
+}
+
