+++ src/domain/index.js
@@ -1,0 +1,6 @@
+
+const News = require('./news')
+
+module.exports = {
+    News
+}
