+++ src/interfaces/http/middlewares/error_handler.js
@@ -1,10 +1,9 @@
 const Joi = require('joi');
-
-const validateNews = (newsData) => {
+const error_handler = (newsData) => {
     const schema = {
         name: Joi.string().min(3).required()
     };
     return Joi.validate(newsData, schema);
 }
 
-module.exports = validateNews
+module.exports = error_handler
